Don't actually execute the git command via hooks
Because of the way plushook does buffering, we can't really hook
the git command up to it. As such, we just use the hook for
notification that the command is about to run instead (which is
enough for the one use case the hook has right now).

This *sort* of contradicts what I was thinking toward the end of #8.