Run Git command in the hook, not after it
Plugins handling their own hooks wasn't an established pattern when
I originally wrote this plugin, which is the main reason it didn't
work this way originally. These days, pretty much all commands are
implemented in a hook which is called by the command, so this is
really the only sensible way for this to be handled.